backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
redis==5.1.1
django-redis==5.4.0
anthropic==0.72.0
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.4.0
PyYAML==6.0.2
